+++ Lab06/Collider.h
@@ -4,7 +4,7 @@
 class Collider : public Actor
 {
 public:
-	Collider(class Game* game, int width, int height);
+	Collider(class Game* game, float width, float height);
 	class CollisionComponent* GetCollisionComponent() { return mCollisionComponent; };
 
 protected:
